requirements.txt
catboost==0.24.4
category-encoders==2.2.2
libzero==0.0.3.dev7
numba==0.53.1
optuna==2.6.0
pandas==1.2.3
pynvml==8.0.4
pytomlpp==0.3.5
scikit-learn==0.24.1
scipy==1.6.1
tensorboard==2.4.1
tqdm==4.59.0
xgboost==1.3.3

# Tools
black
flake8
icecream
isort

# Jupyter
ipywidgets
jupyterlab
jupyterlab-nvdashboard
voila

# Visualization
bokeh
colorcet
holoviews
matplotlib
panel
seaborn
